+++ src/app/create-ad/(formGroup)/preview/_components/image-grid.tsx
