+++ src/app/create-ad/(formGroup)/preview/_components/preview-content.tsx
