+++ src/app/create-ad/(formGroup)/preview/_components/selected-image-preview.tsx
