+++ src/app/create-ad/(formGroup)/preview/page.tsx
@@ -1,0 +1,9 @@
+import PreviewContent from "./_components/preview-content";
+
+export default function Page() {
+  return (
+      <section className="flex flex-col items-center justify-center gap-8 w-full max-w-[879px] mx-auto bg-white rounded-[20px]">
+        <PreviewContent />
+      </section>
+  );
+}
